float handling

+++ variadic_functions/3-print_all.c
@@ -58,7 +58,7 @@ void printInt(va_list intList)
  */
 void printFlot(va_list flotList)
 {
-	printf("%f", (float)va_arg(flotList, int));
+	printf("%f", va_arg(flotList, double));
 }
 /**
  * printStr - prints the Strings
